upd getAll meth
