--- controllers/contacts/getAll.js
@@ -1,13 +1,0 @@
-const contactsService = require("../../models/contacts")
-
-const getAll = async (req, res, next) => {
-	try {
-		const contacts = await contactsService.listContacts();
-		
-		res.json(contacts);
-	} catch (error) {
-		next(error);
-	}
-};
-
-module.exports = getAll;
