--- controllers/contacts/index.js
@@ -1,13 +1,0 @@
-const add = require("./add");
-const deleteById = require("./deleteById");
-const getAll = require("./getAll");
-const getById = require("./getById");
-const updById = require("./updById");
-
-module.exports = {
-	add,
-	deleteById,
-	getAll,
-	getById,
-	updById,
-}
