--- controllers/index.js
@@ -1,5 +1,0 @@
-const contactsCtrl = require("./contacts")
-
-module.exports = {
-	contactsCtrl,
-};
