--- models/contacts.json
@@ -1,56 +1,0 @@
-[
-  {
-    "id": "vza2RIzNGIwutCVCs4mCL",
-    "name": "Serj",
-    "email": "Nsv",
-    "phone": "(000) 6666"
-  },
-  {
-    "id": "05olLMgyVQdWRwgKfg5J6",
-    "name": "Cyrus Jackson",
-    "email": "nibh@semsempererat.com",
-    "phone": "(501) 472-5218"
-  },
-  {
-    "id": "1DEXoP8AuCGYc1YgoQ6hw",
-    "name": "Abbot Franks",
-    "email": "scelerisque@magnis.org",
-    "phone": "(186) 568-3720"
-  },
-  {
-    "id": "Z5sbDlS7pCzNsnAHLtDJd",
-    "name": "Reuben Henry",
-    "email": "pharetra.ut@dictum.co.uk",
-    "phone": "(715) 598-5792"
-  },
-  {
-    "id": "C9sjBfCo4UJCWjzBnOtxl",
-    "name": "Simon Morton",
-    "email": "dui.Fusce.diam@Donec.com",
-    "phone": "(233) 738-2360"
-  },
-  {
-    "id": "e6ywwRe4jcqxXfCZOj_1e",
-    "name": "Thomas Lucas",
-    "email": "nec@Nulla.com",
-    "phone": "(704) 398-7993"
-  },
-  {
-    "id": "cg9W7iNwQ",
-    "name": "Serj",
-    "email": "Nsv",
-    "phone": "(000) 61632"
-  },
-  {
-    "id": "opq3r94u8",
-    "name": "Serj",
-    "email": "Nsv",
-    "phone": "(000) 2232"
-  },
-  {
-    "id": "ztkVXWDKq",
-    "name": "Serj",
-    "email": "Nsv",
-    "phone": "(000) 3217"
-  }
-]
